Use defusedxml instead of etree
Use defusedxml instead of etree. This mitigates the possible
vulnerability to billion-laughs and quadratic blowup. The following
exceptions may be thrown when parsing xmlbombs.

 - EntitiesForbidden: when parsing xml with a ENTITY declaration
 - DTDForbidden: when parsing xml with a DOCTYPE processing instruction

The xml we expect should not have a DOCTYPE processing instruction or
ENTITY declaration. As a result, there should be no side effects due to
this change under normal operation.